Fix shared volume, fix secure websocket in javascript (client-side)
For the volumes, the nginx and rover container share the `app`
directory. The reason for this is that in the future, I might start to
use nginx to serve static content.

For the websocket to work in browser, I just had to add to change `ws`
into `wss`. Port is not relevant, since 80 will be redirected to 443
anyway.